move test fixtures
