+++ support-frontend/assets/pages/digital-subscription-landing/components/paymentSelection/__tests__/__fixtures__/productPrices.ts
