+++ support-frontend/assets/pages/digital-subscription-landing/components/paymentSelection/__tests__/paymentSelectionTests.ts
@@ -5,7 +5,7 @@ import {
 	getProductPrice,
 	getSavingPercentage,
 } from '../helpers/paymentSelection';
-import { productPrices } from './fixtures/productPrices';
+import { productPrices } from './__fixtures__/productPrices';
 
 jest.mock('ophan');
