+++ support-frontend/ts-error-history.csv
@@ -180,4 +180,19 @@
 1650364927000,195,4f10e278cc56eee58b27aea61a7cb54b9359c151
 1650890324000,195,ce7fd3c12b01c057b3ccd4ebc88b07407eda3613
 1650965553000,188,b720479cc39087a85a58ffeabc35f4fc88ea89dd
-1651672024000,210,HEAD
+1651145885000,188,34a01860ac4ad467d7a471528d6f0439324e21c6
+1651151984000,200,20ca383fe0d560dab07f2fbc8ce562a3f3ef3d8e
+1651154534000,188,c0e669a2d916ae9a07324766c2cc3f19757e2f73
+1651158775000,200,6c294edbfcc1bc1937d79b76777d0a2fb959ddef
+1651224842000,177,43302e7d37facb08a840de2294e2dc84e448a14e
+1651232414000,177,cbc4f9728883b3fcd0c0ce3914fb11e0e0360971
+1651239629000,177,541f11b3e3bf894218b7717a849d9993fb53891a
+1651245211000,177,9741a56132a9589277720dd2a2d089fc9bb3a6e4
+1651578250000,177,e74d2907fb13954e662b9eced8d8c2e342c0f624
+1651589267000,177,102331367b2aba6cddc79faf372f12d87eecbce1
+1651591781000,166,3fda79a8494cba122689f9e6840368b7d64e4dbd
+1651656376000,166,53931d00ea2d6abad69787910b3173969b7db6da
+1651672633000,162,2478c486f7a7e04f763c358ea120ddea9437230b
+1651741698000,162,fb9cef42d5d1feac04be860f50a14cf283526572
+1651746399000,162,991f15d0508cd77e69ccfbf57977985bd2e9f64e
+1651836159000,184,HEAD
